Added lockfree support for 64Bit Windows
git-svn-id: https://svn.cct.lsu.edu/repos/projects/parallex/trunk/hpx@711 5079c8e4-9419-0410-93bb-9e5ee278c886

+++ src/runtime/naming/server/request_handler.cpp
@@ -437,7 +437,7 @@ namespace hpx { namespace naming { namespace server
 #if BOOST_VERSION < 103600
     double extract_count(std::pair<double, std::size_t> const& p)
     {
-        return p.second;
+        return double(p.second);
     }
 #endif
